Add files via upload
Description:
CNC machining is a manufacturing process in which programmable computer software codes
dictates the movement of factory tools and machinery. The process can be used to control a
range of complex machinery, from grinders and lathes to mills and routers. With CNC
machining, three-dimensional cutting/writing/3d parts making tasks can be accomplished in a
single set of prompts.
Short for “computer numerical control,” the CNC process runs in contrast to — and thereby
supersedes — the limitations of manual control, where live operators are needed to prompt and
guide the commands of machining tools via levers, buttons and wheels. To the onlooker, a
CNC system might resemble a regular set of computer components, but the software programs
and consoles employed in CNC machining distinguish it from all other forms of computation
In CNC, machines are operated via numerical control, wherein a software program is
designated to control an object. The language behind CNC machining is alternately referred to
as G-code, and it’s written to control the vario